requirements.txt
albumentations==0.5.2
segmentation-models-pytorch==0.1.3
matplotlib==3.4.2
matplotlib-inline==0.1.2
numpy==1.20.3
opencv-python==4.5.2.52
opencv-python-headless==4.5.2.52
pandas==1.2.4
scikit-image==0.18.1
scikit-learn==0.24.2
torch==1.8.1
torchvision==0.9.1
wandb==0.10.31
Pillow==8.3.0